pyproject.toml
[tool.poetry]
name = "pic"
version = "0.1.0"
description = ""
authors = ["David Moya <3nueves@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
pyyaml = "^6.0.2"
gitpython = "^3.1.43"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
